+++ releases.md
@@ -1,5 +1,10 @@
 # Releases
 
+## Unreleased
+
+  - Support grouped IP and Unix-domain-socket addresses in EDS endpoints.
+  - Support selecting HTTP/1 or HTTP/2 for generated clusters.
+
 ## v0.1.0
 
   - Initial extraction from `async-grpc`.